Module Avdevice

Tits module contains input and output devices for grabbing from and rendering to many common multimedia input/output software frameworks.

val init : unit -> unit

Initialize the module. This is done implicitely if you use any of the module's API but is here to provide an easy way to make sure that the module is explicitely linked by the OCaml compiler. NOT thread-safe!

val get_audio_input_formats : unit -> (Avutil.input, Avutil.audio) Avutil.format list

Return the audio input devices formats.

val get_default_audio_input_format : unit -> (Avutil.input, Avutil.audio) Avutil.format

Return the default audio input device format.

val get_video_input_formats : unit -> (Avutil.input, Avutil.video) Avutil.format list

Return the video input devices formats.

val get_default_video_input_format : unit -> (Avutil.input, Avutil.video) Avutil.format

Return the default video input device format.

val get_audio_output_formats : unit -> (Avutil.output, Avutil.audio) Avutil.format list

Return the audio output devices formats.

val get_default_audio_output_format : unit -> (Avutil.output, Avutil.audio) Avutil.format

Return the default audio output device format.

val get_video_output_formats : unit -> (Avutil.output, Avutil.video) Avutil.format list

Return the video output devices formats.

val get_default_video_output_format : unit -> (Avutil.output, Avutil.video) Avutil.format

Return the default video output device format.

val open_audio_input : string -> Avutil.input Avutil.container

Open the audio input device from its name. Raise Error if the device is not found.

val open_default_audio_input : unit -> Avutil.input Avutil.container

Open the default audio input device from its name. Raise Error if the device is not found.

val open_video_input : string -> Avutil.input Avutil.container

Open the video input device from its name. Raise Error if the device is not found.

val open_default_video_input : unit -> Avutil.input Avutil.container

Open the default video input device from its name. Raise Error if the device is not found.

val open_audio_output : ?⁠opts:Avutil.opts -> string -> Avutil.output Avutil.container

Open the audio output device from its name. Raise Error if the device is not found.

val open_default_audio_output : ?⁠opts:Avutil.opts -> unit -> Avutil.output Avutil.container

Open the default audio output device from its name. Raise Error if the device is not found.

val open_video_output : ?⁠opts:Avutil.opts -> string -> Avutil.output Avutil.container

Open the video output device from its name. Raise Error if the device is not found.

val open_default_video_output : ?⁠opts:Avutil.opts -> unit -> Avutil.output Avutil.container

Open the default video output device from its name. Raise Error if the device is not found.

module App_to_dev : sig ... end

Application to device communication

module Dev_to_app : sig ... end

Device to application communication